YT-24426: Fix optional date
* Changelog entry
  Type: fix
  Component: proxy

  Fixing the reading of tables with date type columns in arrow format.
commit_hash:2b6c31c267c7bc3b3dba247387edce1713d3625a
