+++ yt/yt/client/table_client/columnar-inl.h
@@ -4,6 +4,8 @@
 #include "columnar.h"
 #endif
 
+#include "helpers.h"
+
 #include <library/cpp/yt/coding/zig_zag.h>
 
 namespace NYT::NTableClient {
@@ -67,6 +69,7 @@ void DecodeVectorRleImpl(
     ui64 baseValue,
     TRange<ui32> dictionaryIndexes,
     TRange<ui64> rleIndexes,
+    TRef bitmap,
     TGetter getter,
     TConsumer consumer)
 {
@@ -81,17 +84,30 @@ void DecodeVectorRleImpl(
                 break;
             }
             decltype(getter(0)) currentValue;
+            bool isNull = false;
             if constexpr(WithDictionary) {
                 auto dictionaryIndex = dictionaryIndexes[currentRleIndex];
                 if (dictionaryIndex == 0) {
                     currentValue = {};
                 } else {
-                    currentValue = getter(dictionaryIndex - 1);
+                    if (bitmap && GetBit(bitmap, dictionaryIndex - 1)) {
+                        isNull = true;
+                    } else {
+                        currentValue = getter(dictionaryIndex - 1);
+                    }
+
                 }
             } else {
-                currentValue = getter(currentRleIndex);
+                if (bitmap && GetBit(bitmap, currentRleIndex)) {
+                    isNull = true;
+                } else {
+                    currentValue = getter(currentRleIndex);
+                }
             }
-            currentDecodedValue = TValueDecoder<WithBaseValue, WithZigZag, T>::Run(currentValue, baseValue);
+            currentDecodedValue = {};
+            if (!isNull) {
+                currentDecodedValue = TValueDecoder<WithBaseValue, WithZigZag, T>::Run(currentValue, baseValue);
+            }
             ++currentRleIndex;
             thresholdIndex = currentRleIndex < std::ssize(rleIndexes)
                 ? std::min(static_cast<i64>(rleIndexes[currentRleIndex]), endIndex)
@@ -131,22 +147,36 @@ void DecodeVectorDirectImpl(
     i64 endIndex,
     ui64 baseValue,
     TRange<ui32> dictionaryIndexes,
+    TRef bitmap,
     TGetter getter,
     TConsumer consumer)
 {
     for (i64 index = startIndex; index < endIndex; ++index) {
+        bool isNull = false;
         decltype(getter(0)) value;
         if constexpr(WithDictionary) {
             auto dictionaryIndex = dictionaryIndexes[index];
             if (dictionaryIndex == 0) {
                 value = {};
             } else {
-                value = getter(dictionaryIndex - 1);
+                if (bitmap && GetBit(bitmap, dictionaryIndex - 1)) {
+                    isNull = true;
+                } else {
+                    value = getter(dictionaryIndex - 1);
+                }
             }
         } else {
-            value = getter(index);
+            if (bitmap && GetBit(bitmap, index)) {
+                isNull = true;
+            } else {
+                value = getter(index);
+            }
         }
-        auto decodedValue = TValueDecoder<WithBaseValue, WithZigZag, T>::Run(value, baseValue);
+        decltype(getter(0)) decodedValue = {};
+        if (!isNull) {
+            decodedValue = TValueDecoder<WithBaseValue, WithZigZag, T>::Run(value, baseValue);
+        }
+
         consumer(decodedValue);
     }
 }
@@ -172,6 +202,7 @@ void DecodeVectorImpl(
     ui64 baseValue,
     TRange<ui32> dictionaryIndexes,
     TRange<ui64> rleIndexes,
+    TRef bitmap,
     TGetter getter,
     TConsumer consumer)
 {
@@ -182,6 +213,7 @@ void DecodeVectorImpl(
             baseValue,
             dictionaryIndexes,
             rleIndexes,
+            bitmap,
             std::forward<TGetter>(getter),
             std::forward<TConsumer>(consumer));
     } else {
@@ -190,6 +222,7 @@ void DecodeVectorImpl(
             endIndex,
             baseValue,
             dictionaryIndexes,
+            bitmap,
             std::forward<TGetter>(getter),
             std::forward<TConsumer>(consumer));
     }
@@ -258,6 +291,7 @@ void DecodeVector(
     bool zigZagEncoded,
     TRange<ui32> dictionaryIndexes,
     TRange<ui64> rleIndexes,
+    TRef bitmap,
     TGetter getter,
     TConsumer consumer)
 {
@@ -271,6 +305,7 @@ void DecodeVector(
             baseValue, \
             dictionaryIndexes, \
             rleIndexes, \
+            bitmap, \
             std::forward<TGetter>(getter), \
             std::forward<TConsumer>(consumer));
 
@@ -324,6 +359,7 @@ void DecodeIntegerVector(
     bool zigZagEncoded,
     TRange<ui32> dictionaryIndexes,
     TRange<ui64> rleIndexes,
+    TRef bitmap,
     TFetcher fetcher,
     TConsumer consumer)
 {
@@ -334,6 +370,7 @@ void DecodeIntegerVector(
         zigZagEncoded,
         dictionaryIndexes,
         rleIndexes,
+        bitmap,
         std::forward<TFetcher>(fetcher),
         std::forward<TConsumer>(consumer));
 }
@@ -358,6 +395,7 @@ void DecodeRawVector(
         false,
         dictionaryIndexes,
         rleIndexes,
+        /*bitmap*/ {},
         std::forward<TFetcher>(fetcher),
         std::forward<TConsumer>(consumer));
 }
